12 спринт Домашнее задание Исправления
