+++ app/src/main/java/com/example/playlistmaker/TrackAdapter.kt
@@ -1,18 +1,12 @@
 package com.example.playlistmaker
 
+import TrackViewHolder
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import java.text.SimpleDateFormat
-import java.util.Locale
 
 class TrackAdapter(private var trackList: List<Track>,val trackListener: TrackListener):
-    RecyclerView.Adapter<TrackAdapter.TrackViewHolder>() {
+    RecyclerView.Adapter<TrackViewHolder>() {
 
 
 fun updateTracks(newTracks: List<Track>) {
@@ -39,25 +33,6 @@ fun updateTracks(newTracks: List<Track>) {
 
     override fun getItemCount(): Int = trackList.size
 
-
-    class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        private val trackName: TextView = itemView.findViewById(R.id.trackName)
-        private val artistName: TextView = itemView.findViewById(R.id.artistName)
-        private val trackTime: TextView = itemView.findViewById(R.id.trackTime)
-        private val trackImage: ImageView = itemView.findViewById(R.id.trackImage)
-
-        fun bind(track: Track) {
-            trackName.text = track.trackName
-            artistName.text = track.artistName
-            trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(track.trackTime.toLong())
-            Glide.with(itemView)
-                .load(track.artworkUrl100)
-                .centerCrop()
-                .placeholder(R.drawable.placeholder)
-                .transform(RoundedCorners(2))
-                .into(trackImage)
-        }
-    }
     fun interface TrackListener {
         fun onTrackClick(track: Track)
     }
