+++ app/src/main/java/com/example/playlistmaker/TrackViewHolder.kt
@@ -1,0 +1,29 @@
+import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmaker.R
+import com.example.playlistmaker.Track
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    private val trackName: TextView = itemView.findViewById(R.id.trackName)
+    private val artistName: TextView = itemView.findViewById(R.id.artistName)
+    private val trackTime: TextView = itemView.findViewById(R.id.trackTime)
+    private val trackImage: ImageView = itemView.findViewById(R.id.trackImage)
+
+    fun bind(track: Track) {
+        trackName.text = track.trackName
+        artistName.text = track.artistName
+        trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(track.trackTime.toLong())
+        Glide.with(itemView)
+            .load(track.artworkUrl100)
+            .centerCrop()
+            .placeholder(R.drawable.placeholder)
+            .transform(RoundedCorners(2))
+            .into(trackImage)
+    }
+}
