+++ app/src/main/res/drawable/switch_thumb_selector.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/blue"
+        android:state_checked="true" />
+    <item android:color="@color/dark_gray"/>
+</selector>
