+++ app/src/main/res/drawable/switch_track_selector.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="#7A9FBBF3"
+        android:state_checked="true" />
+    <item android:color="#7AAEAFB4"/>
+</selector>
