+++ app/src/main/res/layout/activity_search.xml
@@ -40,8 +40,8 @@
         <FrameLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:paddingLeft="20dp"
-            android:paddingRight="20dp"
+            android:paddingLeft="16dp"
+            android:paddingRight="16dp"
             android:paddingTop="5dp"
             android:layout_gravity="center">
 
@@ -76,7 +76,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="@dimen/top_padding"
             app:layout_constraintTop_toBottomOf="@id/editText"
-            android:padding="@dimen/padding"/>
+            android:padding="@dimen/recyclePadding"/>
 
         <ImageView
             style="@style/NoResultImage"
@@ -145,7 +145,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginTop="8dp"
                     android:visibility="gone"
-                    android:padding="@dimen/padding"/>
+                    android:padding="@dimen/recyclePadding"/>
 
                 <Button
                     style="@style/ClearButtonStyle"
