+++ app/src/main/res/values-night/themes.xml
@@ -53,6 +53,8 @@
         <item name="android:textAllCaps">false</item>
         <item name="android:textSize">@dimen/small_text_size</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
+        <item name="thumbTint">@drawable/switch_thumb_selector</item>
+        <item name="trackTint">@drawable/switch_track_selector</item>
 
     </style>
