+++ app/src/main/res/values/styles.xml
@@ -17,7 +17,7 @@
 
     </style>
 
-    <style name="SettingTextView">
+    <style name="SettingTextView" parent="Widget.MaterialComponents.CompoundButton.Switch">
         <item name="android:layout_height">wrap_content</item>
         <item name="android:layout_gravity">start</item>
         <item name="android:backgroundTint">@color/white</item>
@@ -25,6 +25,9 @@
         <item name="android:textAllCaps">false</item>
         <item name="android:textSize">@dimen/small_text_size</item>
         <item name="android:fontFamily">@font/ys_display_medium</item>
+        <item name="thumbTint">@drawable/switch_thumb_selector</item>
+        <item name="trackTint">@drawable/switch_track_selector</item>
+
 
     </style>
     <style name="PageMock">
@@ -141,12 +144,6 @@
         <item name="android:textAllCaps">false</item>
         <item name="android:shadowRadius">0</item>
     </style>
-
-
-
-
-
-
 </resources>
